Started to setup some basic email parsing
Need to turn this into some real code now.  This was just a spike to
ensure that i can parse email from a gmail account from sparrow/mail on
the iPad and mail on the iPad when used from the Zite application.

Seems to work ok right now to get the first email.  Need to setup some
more tests from different applications.  Currently it does not actually
download the message from gmail but that will have to change eventually.