Created a simple daemon that will monitor for urls
This bot will monitor the jetty@ouelong.com address and pull out urls
from the messages there and add them as content.  Right now it just runs
and checks that email account.  Still need to update the code to
actually add them to the mongoid database.  This will likely require a
"user" object so we can have different libraries.

woo